Fix chasquid compilation
chasquid requires git to be installed, as well as the git repo
to compile to get the version or whatever. This means that the
source must be acquired via a git clone instead of simply
downloading an archive without all the git history. Also, git
does not keep track of date modified so the file integrity
changes each time chasquid is built, so git rev-parse is used
for integrity checking. Hopefully this is cryptographically strong
or something idk